+++ packages/editor-common/web/src/index.js
@@ -46,6 +46,7 @@ export { default as TooltipHost } from './Components/TooltipHost';
 export { default as ColorPicker } from './Components/ColorPicker/ColorPicker';
 export { default as CustomColorPicker } from './Components/ColorPicker/CustomColorPicker';
 export { default as Loader } from './Components/Loader';
+export { default as ErrorMsgWithIcon } from './Components/ErrorMsgWithIcon';
 
 //Modals
 export { default as EditorModals } from './Modals/EditorModals';